+++ src/Controllers/Traits/Crud/Index.php
@@ -287,7 +287,7 @@ trait Index
      * @param array $data
      * @return Response|array
      */
-    protected function trashedIndexResponse(array $data)
+    protected function trashedIndexResponse(LengthAwarePaginator | array $data)
     {
         return $this->indexResponse($data);
     }